Initial create / iconCreator access
I'm questioning if I need this or not, but it successfully serves as an intermediate module. The main purpose of creating this was to have a module that could make elements of the site with the classes and other settings all ready. Let's see if it proves to be useful.